requirements.txt
arm-pyart @ file:///Users/runner/miniforge3/conda-bld/arm_pyart_1637022925381/work
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cftime @ file:///Users/runner/miniforge3/conda-bld/cftime_1642887522786/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1643090971210/work
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1642514170470/work
jmespath==0.10.0
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1635836828404/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1639359003068/work
munkres==1.1.4
netCDF4 @ file:///Users/runner/miniforge3/conda-bld/netcdf4_1636303090422/work
nexradaws==1.1
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1643009311469/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.0
Pillow==8.4.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1642753572664/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1633452062248/work
s3transfer==0.5.0
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1637806969526/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1638334978229/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1640031259003/work
urllib3==1.26.8
xarray @ file:///home/conda/feedstock_root/build_artifacts/xarray_1643438115178/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1641731510635/work
arm-pyart @ file:///Users/runner/miniforge3/conda-bld/arm_pyart_1637022925381/work
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cftime @ file:///Users/runner/miniforge3/conda-bld/cftime_1642887522786/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1643090971210/work
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1642514170470/work
jmespath==0.10.0
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1635836828404/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1639359003068/work
munkres==1.1.4
netCDF4 @ file:///Users/runner/miniforge3/conda-bld/netcdf4_1636303090422/work
nexradaws==1.1
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1643009311469/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.0
Pillow==8.4.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1642753572664/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1633452062248/work
s3transfer==0.5.0
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1637806969526/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1638334978229/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1640031259003/work
urllib3==1.26.8
xarray @ file:///home/conda/feedstock_root/build_artifacts/xarray_1643438115178/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1641731510635/work
arm-pyart @ file:///Users/runner/miniforge3/conda-bld/arm_pyart_1637022925381/work
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cftime @ file:///Users/runner/miniforge3/conda-bld/cftime_1642887522786/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1643090971210/work
grpcio==1.43.0
grpcio-tools==1.43.0
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1642514170470/work
jmespath==0.10.0
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1635836828404/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1639359003068/work
munkres==1.1.4
netCDF4 @ file:///Users/runner/miniforge3/conda-bld/netcdf4_1636303090422/work
nexradaws==1.1
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1643009311469/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.0
Pillow==8.4.0
protobuf==3.19.4
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1642753572664/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1633452062248/work
s3transfer==0.5.0
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1637806969526/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1638334978229/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1640031259003/work
urllib3==1.26.8
xarray @ file:///home/conda/feedstock_root/build_artifacts/xarray_1643438115178/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1641731510635/work
